Moved to SFML 3.0 and can be compiled, but there are some graphical errors that needs to be fixed.

+++ Source/Command.cpp
@@ -2,7 +2,7 @@
 
 Command::Command() :
     action(),
-    category(static_cast<unsigned int>(Category::None))
+    category(Category::None)
 {
 
 }
